refactor: move `migrateLegacyPrType` to `src/types.ts`
Co-locates `migrateLegacyPrType` and `LEGACY_PR_TYPE_MAP` with the
analogous `migrateLegacySeverity` migration in `src/types.ts` for
consistency.

Addresses manki-review review comment on PR #750
https://github.com/manki-review/manki/pull/750#discussion_r3273239498
